requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
